test: split REQ-AU-06 /v1/me integration tests into integration_me_tests.rs
services/control-api/tests/integration_tests.rs grew to 650 lines after
adding the 5 GET /v1/me real-DB tests, exceeding the 600-line CI cap.
Extract the new tests + signup_and_login helper into a sibling
integration_me_tests.rs (Cargo treats each file in tests/ as its own
test binary, so it gets its own preamble — real_db_state, json_body,
collect_body — duplicated by design).

Same pattern used previously for splitting login tests.

- integration_tests.rs: 650 → 367 lines
- integration_me_tests.rs: new, 352 lines
- cargo test -p control-api --lib: 68 passed, 0 failed
- cargo clippy -p control-api --lib --tests -- -D warnings: clean

Addresses CTO review feedback on PR #42.
